requirements.txt
# Runtime dependencies for the ANPR app on modern Python.
# Older pinned packages from the original repo were removed because they no
# longer publish wheels for current Python releases.

torch>=2.2,<2.12
torchvision>=0.17,<0.22
numpy>=1.26,<3
opencv-python-headless>=4.8,<5
Pillow>=10,<12
PyYAML>=6,<7
requests>=2.31,<3
psutil>=5.9,<8
tqdm>=4.66,<5
matplotlib>=3.8,<4
pandas>=2.2,<3
seaborn>=0.13,<0.14
scipy>=1.11,<2
Flask>=3.0,<4
easyocr>=1.7,<2
pytesseract>=0.3.10,<0.4
streamlit>=1.37,<2
